thebox/services/requirements.txt
pyyaml
couchdb
cloudant
dataclasses
jsonpickle
docker
confluent-kafka
flask
flask.ext
flask-CORS
flask-restful-swagger
flask-marshmallow
object-mapper
wheel
numpy
onnxruntime
tensorflow==1.14.0
pymaybe
marshmallow==2.19.5
marshmallow-dataclass==0.6.6
